refs #621 refactor: Divide actions
