+++ assets/js/components/projects/lists.tsx
@@ -3,7 +3,8 @@ import { Button } from 'react-bootstrap'
 import { ThunkDispatch } from 'redux-thunk'
 import { Link, RouteComponentProps } from 'react-router-dom'
 
-import Actions, { getLists, getProject, openDelete, closeDelete, openNewList, closeNewList, openEditProject, closeEditProject } from '@/actions/projects/lists'
+import Actions, { getLists, openDelete, closeDelete, openNewList, closeNewList, openEditProject, closeEditProject } from '@/actions/projects/lists'
+import { getProject } from '@/actions/projects/show'
 import { RootStore } from '@/reducers/index'
 import List from './lists/list.tsx'
 import styles from './lists.scss'
